requirements.txt
flask==3.1.0
flask-cors==5.0.0
google-generativeai==0.8.5
langchain==0.3.27
langchain-community==0.3.27
faiss-cpu==1.12.0
python-dotenv==1.1.1
pdfplumber==0.11.7
PyPDF2==3.0.1
langdetect==1.0.9